+++ internal/backend/local/doc.go
@@ -1,0 +1,7 @@
+// Copyright (c) 2026 Steve Taranto <staranto@gmail.com>.
+// SPDX-License-Identifier: Apache-2.0
+// no-cloc
+
+// Package cloud implements backend integrations for local and exposes common
+// behaviors for querying runs, state, and state versions.
+package local