+++ internal/backend/s3/doc.go
@@ -1,0 +1,7 @@
+// Copyright (c) 2026 Steve Taranto <staranto@gmail.com>.
+// SPDX-License-Identifier: Apache-2.0
+// no-cloc
+
+// Package s3 implements tfctl's Terraform backend support for S3-backed
+// Terraform state.
+package s3