+++ internal/cacheutil/doc.go
@@ -1,0 +1,6 @@
+// Copyright (c) 2026 Steve Taranto <staranto@gmail.com>.
+// SPDX-License-Identifier: Apache-2.0
+// no-cloc
+
+// Package cacheutil provides a small, on-disk cache used by tfctl.
+package cacheutil