+++ internal/command/si/doc.go
@@ -1,0 +1,7 @@
+// Copyright (c) 2026 Steve Taranto <staranto@gmail.com>.
+// SPDX-License-Identifier: Apache-2.0
+// no-cloc
+
+// Package si is a dormant experiment to provide an interactive state query with
+// more functionality than `terraform console`.
+package si