+++ internal/hungarian/doc.go
@@ -1,0 +1,6 @@
+// Copyright (c) 2026 Steve Taranto <staranto@gmail.com>.
+// SPDX-License-Identifier: Apache-2.0
+// no-cloc
+
+// Package hungarian detects Hungarian-style naming in Terraform resource names.
+package hungarian